[FLINK-5481] [core] Followups to the typeinfo/Types utility   - Move new 'Types' class to 'common' scope, not 'java api'.   - Make the method names the same in Scala and Java
